added SectionHeader, deleted unused comps
